refactor(byok): remove direct-slot workflow
Keep BYOK on the shared graph-slot backend through family Recipes and manual graph selection. Remove the separate kernel YAML CLI and family slot registry, and align the Qwen FlashInfer POC with the attention-only Recipe ABI.
