--- python/tensorrt_model_connect/families/qwen/examples/byok_flashinfer/prompt.txt
@@ -1,1 +1,0 @@
-Explain grouped-query attention in two short sentences.
