--- python/tensorrt_model_connect/families/qwen/examples/byok_flashinfer/qwen3-flashinfer.yaml.in
@@ -1,12 +1,0 @@
-# SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-# SPDX-License-Identifier: Apache-2.0
-
-schema_version: 1
-slot: qwen.decode_attention@1
-instances:
-  all: true
-  expect_count: 36
-kernel:
-  library: ./flashinfer-qwen3.so
-  sha256: "@SHA256@"
-  function: run
