+++ python/tensorrt_model_connect/families/qwen/kernels/flashinfer_device_kv_length.patch
@@ -1,0 +1,36 @@
+diff --git a/flashinfer/cute_dsl/attention/gqa_decode.py b/flashinfer/cute_dsl/attention/gqa_decode.py
+index 79d35bc..0017b9b 100755
+--- a/flashinfer/cute_dsl/attention/gqa_decode.py
++++ b/flashinfer/cute_dsl/attention/gqa_decode.py
+@@ -208,2 +208,3 @@ class GroupedQueryAttentionDecode:
+         enable_pdl: bool = True,
++        kv_lens: Optional[cute.Tensor] = None,
+     ):
+@@ -242,2 +243,5 @@ class GroupedQueryAttentionDecode:
+             toggle.
++        kv_lens
++            Optional device-resident valid KV length per batch. ``None`` uses
++            the K/V TensorView sequence dimension, preserving existing calls.
+         """
+@@ -479,2 +483,4 @@ class GroupedQueryAttentionDecode:
+             tma_tensor_v,
++            # Runtime sequence lengths
++            kv_lens,
+             # O
+@@ -538,2 +544,3 @@ class GroupedQueryAttentionDecode:
+         mV: cute.Tensor,
++        kv_lens: Optional[cute.Tensor],
+         # O
+@@ -636,3 +643,2 @@ class GroupedQueryAttentionDecode:
+         # Unpack multimodes
+-        seqlen = mK.shape[0]
+         grouped_heads, prediction = mQ.shape[0]
+@@ -644,3 +650,7 @@ class GroupedQueryAttentionDecode:
+         coord_hk, coord_b = coord_hb
+-
++        if cutlass.const_expr(kv_lens is None):
++            seqlen = mK.shape[0]
++        else:
++            seqlen = kv_lens[coord_b]
++
+         # Runtime control flow
