+++ tests/cpp/test_cli_args.cpp
@@ -74,10 +74,6 @@ void test_build_forwards_args_verbatim() {
     check(args.build_args[0] == "Example/Decoder-0.6B", "build forwards model");
     check(args.build_args[4] == "fp16", "build forwards final value");
 
-    auto kernel = parse({"trtmc", "kernel", "slots", "Example/Decoder-0.6B"});
-    check(kernel.command == "kernel" && kernel.build_args.size() == 2,
-          "kernel command forwards to Python");
-
     auto graph = parse({"trtmc", "graph", "list", "graph.json"});
     check(graph.command == "graph" && graph.build_args.size() == 2,
           "graph command forwards to Python");
